Added Taost, Confirmation dialog, Loader and uint teest cases

+++ src/app/members/members.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { AppService } from '../app.service';
 import { Router } from '@angular/router';
+import { NotificationService } from '../notification/notification.service'
+import { ConfirmationDialogService } from '../confirmation-dialog/confirmation-dialog.service';
+import { SpinnerVisibilityService } from 'ng-http-loader';
 
 @Component({
   selector: 'app-members',
@@ -9,18 +12,30 @@ import { Router } from '@angular/router';
 })
 export class MembersComponent implements OnInit {
   members = [];
-
+    
   constructor(
-    public appService: AppService, 
+    public appService: AppService,
+    private notifyService : NotificationService, 
+    private confirmationDialogService: ConfirmationDialogService,
+    private spinner: SpinnerVisibilityService,
     private router: Router) {}
 
   /** 
    * Function fetch all members
   */
   ngOnInit() {
+    // Show Loader
+    this.spinner.show();
+
     // Call fetch all members API and set response in members array
     this.appService.getMembers()
-      .subscribe(members => (this.members = members));
+      .subscribe(members => {
+        // Set Members 
+        this.members = members
+
+        // Hide Loader
+        this.spinner.hide();    
+    });
   }
 
   /** 
@@ -45,15 +60,28 @@ export class MembersComponent implements OnInit {
   /** 
    * Function delete the member
   */
- deleteMemberById(id: number) {
-    // this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
-    // .then((confirmed) => 
-    //   console.log('User confirmed:', confirmed))
-    // .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+ deleteMemberById(id: number) {   
+    this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to delete member ?')
+    .then((confirmed) => {
+      console.log(confirmed);
+      if(confirmed) {
+        // Show Loader
+        this.spinner.show();
 
-    // Call delete member API
-    this.appService.deleteMember(id).subscribe( data => {
-      this.members = this.members.filter(member =>  member.id !== id);
-    });
+        // Call delete member API
+        this.appService.deleteMember(id).subscribe( data => {
+          this.members = this.members.filter(member =>  member.id !== id);
+
+          // Set Notification
+          this.notifyService.showSuccess("Member deleted successfully", "Success!")
+          
+          // Hide Loader
+          this.spinner.hide();    
+        });
+      }
+    })
+    .catch(() => 
+      console.log('User discard to delete.')
+    );
   }
 }
